refactor: reorganize files a bit and update to use api client
