+++ src/commands/check.ts
@@ -1,5 +1,5 @@
 import { CommandInteraction, SlashCommandBuilder } from "discord.js";
-import { ChainPatrolApiClient, AssetType } from "../utils/ChainPatrolApiClient";
+import { ChainPatrolApiClient, AssetType } from "../utils/api";
 import { defangUrl } from "../utils/url";
 
 export const data = new SlashCommandBuilder()
