+++ src/utils/url.ts
@@ -1,3 +1,9 @@
 export function defangUrl(url: string) {
   return url.replace(".", "(dot)");
 }
+
+export const URL_REGEX = /https?:\/\/(?:www\.)?([^\s]+\.[^\s]+)/gi;
+
+export function extractUrls(str: string) {
+  return str.match(URL_REGEX);
+}
